Display ROM file name instead of full path
On some devices, the path of a file cannot be accurately determined. This caused the displayed path of ROMs to be useless, with no actual indication of which file that particular ROM was pointing to. With this change, the name of the ROM's file should always be properly displayed, at the expense of not displaying the full path